requirements.txt
Django==5.0.6
pandas==2.2.2
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
python-dotenv==1.0.1
drf-yasg==1.21.7